requirements.txt


daphne==2.4.0
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.8.1
django-environ==0.4.5
django-filter==2.1.0
django-heroku==0.3.1
django-map-widgets==0.2.2
django-storages==1.7.1
botocore==1.12.186
cement==2.8.2
colorama==0.3.9
jmespath==0.9.4
pathspec==0.5.9

requests==2.20.1
semantic-version==2.5.0

gunicorn==20.1.0


pillow==8.0.1
django_google_maps
folium==0.13.0



google-cloud-storage==2.14.0

